requirements.txt
discord.py
aiosqlite
datetime